Fix filters in ExceptionRaceConditionIT
Unfortunately I messed up all the filters (probably got confused by
the very generic sounding names) and it resulted in us just asserting
on nothing.

I am a bit worried about the fragility of those tests. If you mistype
a choice name you run into the same issue. A potentially more robust
option might be to assert that the filters are non-empty or run them
with the assertion reversed and check that they fail but that seems
out of scope for this PR.

I did reverse all assertions manually and checked that the tests do
fail so they properly assert on what they should.

Very sorry for messing this up

changelog_begin
changelog_end
