+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/infrastructure/RaceConditionTests.scala
@@ -5,8 +5,7 @@ package com.daml.ledger.api.testtool.infrastructure
 
 import com.daml.ledger.api.testtool.infrastructure.Assertions._
 import com.daml.ledger.api.testtool.infrastructure.participant.ParticipantTestContext
-import com.daml.ledger.api.v1.transaction.{TransactionTree, TreeEvent}
-import com.daml.ledger.api.v1.value.RecordField
+import com.daml.ledger.api.v1.transaction.TransactionTree
 import com.daml.ledger.client.binding.Primitive
 import com.daml.lf.data.{Bytes, Ref}
 import com.daml.timer.Delayed
@@ -74,79 +73,5 @@ private[testtool] object RaceConditionTests {
 
   private def offsetBytes(offset: String): Bytes = {
     Bytes.fromHexString(Ref.HexString.assertFromString(offset))
-  }
-
-  object TransactionUtil {
-
-    private implicit class TransactionTreeTestOps(tx: TransactionTree) {
-      def hasEventsNumber(expectedNumberOfEvents: Int): Boolean =
-        tx.eventsById.size == expectedNumberOfEvents
-
-      def containsEvent(condition: TreeEvent => Boolean): Boolean =
-        tx.eventsById.values.toList.exists(condition)
-    }
-
-    private def isCreated(templateName: String)(event: TreeEvent): Boolean =
-      event.kind.isCreated && event.getCreated.templateId.exists(_.entityName == templateName)
-
-    private def isExerciseEvent(choiceName: String)(event: TreeEvent): Boolean =
-      event.kind.isExercised && event.getExercised.choice == choiceName
-
-    def isCreateDummyContract(tx: TransactionTree): Boolean =
-      tx.containsEvent(isCreated(RaceTests.DummyContract.TemplateName))
-
-    def isCreateNonTransient(tx: TransactionTree): Boolean =
-      tx.hasEventsNumber(1) &&
-        tx.containsEvent(isCreated(RaceTests.ContractWithKey.TemplateName))
-
-    def isTransientCreate(tx: TransactionTree): Boolean =
-      tx.containsEvent(isExerciseEvent(RaceTests.CreateWrapper.ChoiceCreateTransient)) &&
-        tx.containsEvent(isCreated(RaceTests.ContractWithKey.TemplateName))
-
-    def isArchival(tx: TransactionTree): Boolean =
-      tx.hasEventsNumber(1) &&
-        tx.containsEvent(isExerciseEvent(RaceTests.ContractWithKey.ChoiceArchive))
-
-    def isNonConsumingExercise(tx: TransactionTree): Boolean =
-      tx.hasEventsNumber(1) &&
-        tx.containsEvent(isExerciseEvent(RaceTests.ContractWithKey.ChoiceExercise))
-
-    def isFetch(tx: TransactionTree): Boolean =
-      tx.hasEventsNumber(1) &&
-        tx.containsEvent(isExerciseEvent(RaceTests.FetchWrapper.ChoiceFetch))
-
-    private def isFoundContractField(found: Boolean)(field: RecordField) = {
-      field.label == "found" && field.value.exists(_.getBool == found)
-    }
-
-    def isContractLookup(success: Boolean)(tx: TransactionTree): Boolean =
-      tx.containsEvent { event =>
-        isCreated(RaceTests.LookupResult.TemplateName)(event) &&
-        event.getCreated.getCreateArguments.fields.exists(isFoundContractField(found = success))
-      }
-  }
-
-  private object RaceTests {
-    object ContractWithKey {
-      val TemplateName = "ContractWithKey"
-      val ChoiceArchive = "ContractWithKey_Archive"
-      val ChoiceExercise = "ContractWithKey_Exercise"
-    }
-
-    object DummyContract {
-      val TemplateName = "DummyContract"
-    }
-
-    object FetchWrapper {
-      val ChoiceFetch = "FetchWrapper_Fetch"
-    }
-
-    object LookupResult {
-      val TemplateName = "LookupResult"
-    }
-
-    object CreateWrapper {
-      val ChoiceCreateTransient = "CreateWrapper_CreateTransient"
-    }
   }
 }
