+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/suites/RaceConditionIT.scala
@@ -8,6 +8,8 @@ import com.daml.ledger.api.testtool.infrastructure.Assertions._
 import com.daml.ledger.api.testtool.infrastructure.LedgerTestSuite
 import com.daml.ledger.api.testtool.infrastructure.participant.ParticipantTestContext
 import com.daml.ledger.api.testtool.infrastructure.RaceConditionTests._
+import com.daml.ledger.api.v1.transaction.{TransactionTree, TreeEvent}
+import com.daml.ledger.api.v1.value.RecordField
 import com.daml.ledger.client.binding.Primitive
 import com.daml.ledger.test.semantic.RaceTests._
 import com.daml.timer.Delayed
@@ -49,7 +51,7 @@ final class RaceConditionIT extends LedgerTestSuite {
       }
       transactions <- transactions(ledger, alice)
     } yield {
-      import TransactionUtil._
+      import RaceConditionIT.TransactionUtil._
       assertSingleton(
         "Successful contract archivals",
         transactions.filter(isArchival),
@@ -85,7 +87,7 @@ final class RaceConditionIT extends LedgerTestSuite {
       ) // Create a dummy contract to ensure that we're not stuck with previous commands
       transactions <- transactions(ledger, alice)
     } yield {
-      import TransactionUtil._
+      import RaceConditionIT.TransactionUtil._
 
       assert(
         isCreateNonTransient(transactions.head),
@@ -130,7 +132,7 @@ final class RaceConditionIT extends LedgerTestSuite {
       )
       transactions <- transactions(ledger, alice)
     } yield {
-      import TransactionUtil._
+      import RaceConditionIT.TransactionUtil._
 
       // We deliberately allow situations where no non-transient contract is created and verify the transactions
       // order when such contract is actually created.
@@ -155,7 +157,7 @@ final class RaceConditionIT extends LedgerTestSuite {
       )
       transactions <- transactions(ledger, alice)
     } yield {
-      import TransactionUtil._
+      import RaceConditionIT.TransactionUtil._
       val archivalTransaction = assertSingleton("archivals", transactions.filter(isArchival))
       transactions
         .filter(isNonConsumingExercise)
@@ -177,7 +179,7 @@ final class RaceConditionIT extends LedgerTestSuite {
       )
       transactions <- transactions(ledger, alice)
     } yield {
-      import TransactionUtil._
+      import RaceConditionIT.TransactionUtil._
       val archivalTransaction = assertSingleton("archivals", transactions.filter(isArchival))
       transactions
         .filter(isFetch)
@@ -199,7 +201,7 @@ final class RaceConditionIT extends LedgerTestSuite {
       )
       transactions <- transactions(ledger, alice)
     } yield {
-      import TransactionUtil._
+      import RaceConditionIT.TransactionUtil._
       val archivalTransaction = assertSingleton("archivals", transactions.filter(isArchival))
       transactions
         .filter(isContractLookup(success = true))
@@ -220,7 +222,7 @@ final class RaceConditionIT extends LedgerTestSuite {
       )
       transactions <- transactions(ledger, alice)
     } yield {
-      import TransactionUtil._
+      import RaceConditionIT.TransactionUtil._
       val createNonTransientTransaction = assertSingleton(
         "create-non-transient transactions",
         transactions.filter(isCreateNonTransient),
@@ -246,4 +248,80 @@ final class RaceConditionIT extends LedgerTestSuite {
     )(implicit ec => { case Participants(Participant(ledger, party)) =>
       testCase(ec)(ledger)(party)
     })
+}
+
+object RaceConditionIT {
+  object TransactionUtil {
+
+    private implicit class TransactionTreeTestOps(tx: TransactionTree) {
+      def hasEventsNumber(expectedNumberOfEvents: Int): Boolean =
+        tx.eventsById.size == expectedNumberOfEvents
+
+      def containsEvent(condition: TreeEvent => Boolean): Boolean =
+        tx.eventsById.values.toList.exists(condition)
+    }
+
+    private def isCreated(templateName: String)(event: TreeEvent): Boolean =
+      event.kind.isCreated && event.getCreated.templateId.exists(_.entityName == templateName)
+
+    private def isExerciseEvent(choiceName: String)(event: TreeEvent): Boolean =
+      event.kind.isExercised && event.getExercised.choice == choiceName
+
+    def isCreateDummyContract(tx: TransactionTree): Boolean =
+      tx.containsEvent(isCreated(RaceTests.DummyContract.TemplateName))
+
+    def isCreateNonTransient(tx: TransactionTree): Boolean =
+      tx.hasEventsNumber(1) &&
+        tx.containsEvent(isCreated(RaceTests.ContractWithKey.TemplateName))
+
+    def isTransientCreate(tx: TransactionTree): Boolean =
+      tx.containsEvent(isExerciseEvent(RaceTests.CreateWrapper.ChoiceCreateTransient)) &&
+        tx.containsEvent(isCreated(RaceTests.ContractWithKey.TemplateName))
+
+    def isArchival(tx: TransactionTree): Boolean =
+      tx.hasEventsNumber(1) &&
+        tx.containsEvent(isExerciseEvent(RaceTests.ContractWithKey.ChoiceArchive))
+
+    def isNonConsumingExercise(tx: TransactionTree): Boolean =
+      tx.hasEventsNumber(1) &&
+        tx.containsEvent(isExerciseEvent(RaceTests.ContractWithKey.ChoiceExercise))
+
+    def isFetch(tx: TransactionTree): Boolean =
+      tx.hasEventsNumber(1) &&
+        tx.containsEvent(isExerciseEvent(RaceTests.FetchWrapper.ChoiceFetch))
+
+    private def isFoundContractField(found: Boolean)(field: RecordField) = {
+      field.label == "found" && field.value.exists(_.getBool == found)
+    }
+
+    def isContractLookup(success: Boolean)(tx: TransactionTree): Boolean =
+      tx.containsEvent { event =>
+        isCreated(RaceTests.LookupResult.TemplateName)(event) &&
+        event.getCreated.getCreateArguments.fields.exists(isFoundContractField(found = success))
+      }
+  }
+
+  private object RaceTests {
+    object ContractWithKey {
+      val TemplateName = "ContractWithKey"
+      val ChoiceArchive = "ContractWithKey_Archive"
+      val ChoiceExercise = "ContractWithKey_Exercise"
+    }
+
+    object DummyContract {
+      val TemplateName = "DummyContract"
+    }
+
+    object FetchWrapper {
+      val ChoiceFetch = "FetchWrapper_Fetch"
+    }
+
+    object LookupResult {
+      val TemplateName = "LookupResult"
+    }
+
+    object CreateWrapper {
+      val ChoiceCreateTransient = "CreateWrapper_CreateTransient"
+    }
+  }
 }
